airflow/requirements.txt
boto3
botocore
bs4
datasets
DateTime==5.5
Deprecated==1.2.14
numpy
pandas
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv
pytz==2024.2
PyYAML
requests
requests-toolbelt
s3transfer
setuptools
apache-airflow==2.10.2
pymupdf==1.24.11
pymupdf4llm==0.0.17
pillow==10.4.0
matplotlib
selenium==4.25.0
webdriver-manager==4.0.2
snowflake-connector-python==3.12.2
snowflake_uuid==0.1.1
num2words